测试通过NopReport的pdf打印，支持图片生成功能，自动拆分大表格

+++ nop-excel/src/main/java/io/nop/excel/util/UnitsHelper.java
@@ -67,7 +67,7 @@ public class UnitsHelper {
      */
     public static final float DEFAULT_CHARACTER_WIDTH = 7.0017f; // 单位为px
 
-    public static final double DEFAULT_CHARACTER_WIDTH_IN_PT = 6.5;
+    public static final float DEFAULT_CHARACTER_WIDTH_IN_PT = 6.5f;
 
     /**
      * Column widths are in fractional characters, this is the EMU equivalent. One character is defined as the widest
